chore: moved all cfd banner mobile text and desktop text in different component
